[BUGFIX] Prevent TcaColPosItems dataProvider for non-item fields
The colPos items are fetched only for several defined config types.
If the colPos field is an input field (e.g. with renderType hidden),
the items aren't processed and the result of the dataProvider is
inconsistent and wrong. Furthermore a PHP warning is thrown as the
databaseRow value isn't an array anymore.